GRID-2365 Add Shacl#in for multiple triples
TODO: Add support for sh:class rdf:List. This does not work yet as the
triples are not correctly queried and changing this requires a lot of
refactoring and changes to the validator. For now it works by declaring
multiple triples of type sh:in with a literal type xsd:string.